VNL_0001/LiveKit_py/pyproject.toml
[project]
name = "livekit-py"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "livekit-agents[elevenlabs,openai,silero,speechmatics,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.1",
]
